[FEAT] : PostItem 컴포넌트 구현으로 컴포넌트 분리
